\name{pgKDEsphere-internal}

\alias{cart_sph}
\alias{hopt}
\alias{KDE}
\alias{KDE_sph_vMF}
\alias{pgKDE_sph_vMF}
\alias{plot_sphkde}
\alias{Rgr}
\title{Internal pgKDEsphere Functions}
\description{
 Internal pgKDEsphere functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }